add even-store

+++ Pacco.Services.Availability/src/Pacco.Services.Availability.Infrastructure/Mongo/Documents/ResourceDocument.cs
@@ -7,7 +7,7 @@ namespace Pacco.Services.Availability.Infrastructure.Mongo.Documents
     public class ResourceDocument : IIdentifiable<Guid>
     {
         public Guid Id { get; set; }
-        public int Version { get; set; }
+        public long Version { get; set; }
         public IEnumerable<string> Tags { get; set; }
         public IEnumerable<ReservationDocument> Reservations { get; set; }
     }
